Split out main contact tests
